refactor: address page-specific component issues and monolithic composition
- Implemented Suspense boundaries and lazy loading in Dashboard for RecentPosts and UpcomingEvents.
- Added standardized isLoading and error states to useBlog and useResources hooks and their respective UI components.
- Eliminated prop-drilling in ResearchDetail and BlogPost by utilizing useParams in decentralized sub-components.
- Decoupled Contact presentation from network logic by extracting ContactForm and ContactSuccess components.
- Verified changes with Playwright E2E smoke tests and visual inspections.
